Add subtoken_position_increment option for dictionary decompounder.
Once the option subtokenPositionIncrement becomes available in the dictionary decompounder in lucene (https://github.com/apache/lucene/pull/12478) this option should be added to elasticsearch too.

The new option will (as described in the linked PR) allow AND match clauses on subtoken. Checkout the description in the linked PR for more detailed explanation.